requirements.txt
flask
chromadb
sentence-transformers
google-genai